The atomspace moved the location of the NumberNode.
So we move it here, also.
