+++ opencog/nlp/learn/notes
@@ -13257,6 +13257,35 @@ line 460 of compute-mi.scm do-one-pair
 
 (psf 'set-pair-freq fa  1.285376022487705e-8)
 
+change make-batch-mi 
+store-pairs
+make-store
+
+store-atom
+SQLAtomStorage::flushStoreQueue
+deleteValuation(const Handle& key, const Handle& atom)
+get_uuid_column_cb
+get_incoming_uuid_cb
+uvec
+
+Handle getAtom(UUID);
+get_atom_values
+
+(use-modules (opencog) (opencog persist) (opencog persist-sql))
+(sql-open "postgres:///opencog_test?user=opencog_tester&password=cheese")
+
+(cog-delete  (Concept "AAA"))
+(define b (Concept "BBB"))
+
+h1 2 3 4 are nodes
+hl hl2 hl3
+delete hl3 then hl2 then hl
+
+wait -- were teh uuids removed? yes they were.
+
+cog-delete-recursive is weierd
+how to find handles....  handle_to_scm won't work.
+
 
 =========================================================================
 OK-ish misses the mark:
@@ -13296,3 +13325,4 @@ Bugs: clean up zh due to above
 Bugs: cleanup embedded daashes in en 
 
 Bugs: get rid of all-atoms in matrix
+
